# ui/src/generated/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Personal World - Worlds",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/rooms/{provider_id}/actions": {
      "get": {
        "operationId": "action_candidates_api_rooms__provider_id__actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Action Candidates Api Rooms  Provider Id  Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Candidates"
      }
    },
    "/api/rooms/{provider_id}/actions/{room_action_id}/adopt": {
      "post": {
        "description": "Make a room's action callable (never exposed to agents, approval always). Sends nothing.",
        "operationId": "adopt_api_rooms__provider_id__actions__room_action_id__adopt_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "room_action_id",
            "required": true,
            "schema": {
              "title": "Room Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Adopt Api Rooms  Provider Id  Actions  Room Action Id  Adopt Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adopt"
      }
    },
    "/api/rooms/{provider_id}/actions/{room_action_id}/answer": {
      "post": {
        "description": "Answer a room need by running its ALREADY-ADOPTED action through the C3 dispatcher (owner, CSRF).\n\nrequest_authorization -> approve -> consume -> dispatch: the same lifecycle as any write, at most\none network attempt. An unadopted action (or one the room no longer offers) is refused; nothing is\never sent without a consumed authorization, and there is no dispatcher-free path to the room.",
        "operationId": "answer_need_api_rooms__provider_id__actions__room_action_id__answer_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "room_action_id",
            "required": true,
            "schema": {
              "title": "Room Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Answer Need Api Rooms  Provider Id  Actions  Room Action Id  Answer Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Need"
      }
    },
    "/api/rooms/{provider_id}/changed": {
      "post": {
        "description": "The room signalled a change (its ``/changed`` ping): invalidate its cached snapshot.\n\nSends nothing to the room and dispatches nothing — the next read refetches. A provider that is\nnot a room0 room is a 404.",
        "operationId": "room_changed_api_rooms__provider_id__changed_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Room Changed Api Rooms  Provider Id  Changed Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Room Changed"
      }
    },
...
  }
}